requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
django==4.2.23
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
psycopg2-binary==2.9.9
pybreaker==1.2.0
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
